// apps/cms/src/shared/api/generated/data-management/schemas/matchesParams.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea CMS Backend API — Data Management subset
 * Filtered for CMS data management Orval codegen.
 * OpenAPI spec version: v9
 */

export type MatchesParams = {
/**
 * 쿼리 파라미터입니다. 요청값 출처: 필터/페이지네이션/선택 행에서 요청값 전달
 */
businessArea?: string;
/**
 * 쿼리 파라미터입니다. 요청값 출처: 필터/페이지네이션/선택 행에서 요청값 전달
 */
educationTarget?: string;
/**
 * 쿼리 파라미터입니다. 요청값 출처: 필터/페이지네이션/선택 행에서 요청값 전달
 */
grade?: string;
};